Some doc cleanup, made `AsyncExt[Task]` the default, and added (currently failing) queue test
