+++ core/src/main/scala/fs2/async/async.scala
@@ -2,10 +2,6 @@ package fs2
 
 import fs2.util.Catchable
 
-
-/**
- * Created by pach on 10/10/15.
- */
 package object async {
 
   /**
@@ -18,6 +14,17 @@ package object async {
   /** Defined as `[[hold]](None, source.map(Some(_)))` */
   def holdOption[F[_]:AsyncExt,A](source: Stream[F, A]): Stream[F, immutable.Signal[F,Option[A]]] =
      immutable.Signal.holdOption(source)
+
+  /** Create an unbounded asynchronous queue. See [[mutable.Queue]] for more documentation. */
+  def unboundedQueue[F[_]:AsyncExt,A]: F[mutable.Queue[F,A]] =
+    mutable.Queue.unbounded[F,A]
+
+  /**
+   * Create an bounded asynchronous queue. Calls to `enqueue1` will wait until the
+   * queue's size is less than `maxSize`. See [[mutable.Queue]] for more documentation.
+   */
+  def boundedQueue[F[_]:AsyncExt,A](maxSize: Int): F[mutable.Queue[F,A]] =
+    mutable.Queue.bounded[F,A](maxSize)
 
   /**
    * Converts a discrete stream to a signal. Returns a single-element stream.
